to-do here
If the output path of Generate_HRUs is set to the input routing product folder, this copy will not work well.
If geojson is existing in the input folder, it will return an error saying the two geojson files are the same file.
Only works when the output folder path is set to be different to the input folder.